dig.go: Split into files by purpose
The dig.go file has grown past readability.
Break it apart into separate files:

- constructor.go defines constructorNode, its method, and the ability
  to call a constructor.
- container.go defines the container type and its related interfaces and
  methods.
- types.go was renamed to inout.go, defining the In and Out types.
- invoke.go defines the Invoke method of the Container type.
- provide.go does the same for the provide method of the Container type.
- stringer.go was deleted in favor of string methods next to their
  types.

This change contains no behavioral changes. Code has only been moved.
